Naniesiono poprawki do struktury.
