+++ wks-recruiter/src/main/java/pl/lodz/p/it/wks/wksrecruiter/WksRecruiterApplication.java
@@ -6,7 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class WksRecruiterApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(WksRecruiterApplication.class, args);
+    public static void main(String[] args) {
+        SpringApplication.run(WksRecruiterApplication.class, args);
     }
+
 }
