+++ wks-recruiter/src/main/java/pl/lodz/p/it/wks/wksrecruiter/collections/PositionEnum.java
@@ -1,7 +1,9 @@
 package pl.lodz.p.it.wks.wksrecruiter.collections;
 
 public enum PositionEnum {
-    JUNIOR_DEV("Junior Developer"), SENIOR_DEV("Senior Developer"), REPOMAN("Repository Master");
+    JUNIOR_DEV("Junior Developer"),
+    SENIOR_DEV("Senior Developer"),
+    REPOMAN("Repository Master");
     private final String value;
     PositionEnum(final String value) { this.value = value; }
     @Override
